requirements.txt
keras==2.8.0
keras_nightly==2.5.0.dev2021032900
matplotlib==3.5.1
nibabel==3.2.2
numpy==1.21.2
pandas==1.3.5
Pillow==9.1.0
scikit_learn==1.0.2
skimage==0.0
tensorflow==2.8.0
tqdm==4.61.2

Flask==1.1.2